Move all files to `source/` and `test/` directories
