+++ package.json
@@ -12,21 +12,19 @@
 	},
 	"type": "module",
 	"exports": {
-		"types": "./index.d.ts",
-		"default": "./index.js"
+		"types": "./source/index.d.ts",
+		"default": "./source/index.js"
 	},
 	"sideEffects": false,
 	"engines": {
 		"node": ">=18.19"
 	},
 	"scripts": {
-		"test": "xo && ava && tsc index.d.ts"
+		"test": "xo && ava && tsc ./source/index.d.ts"
 	},
 	"files": [
-		"index.js",
-		"index.d.ts",
-		"iterable.js",
-		"windows.js"
+		"source/**/*.js",
+		"source/**/*.d.ts"
 	],
 	"keywords": [
 		"spawn",
