+++ test/fixtures/()[]%0!`.cmd
