+++ test/fixtures/node-flags-path.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 import process from 'node:process';
-import nanoSpawn from '../index.js';
+import nanoSpawn from '../../source/index.js';
 
 await nanoSpawn(process.execPath, ['-p', 'process.execArgv'], {stdout: 'inherit'});
