+++ test/fixtures/node-version.js
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
-import nanoSpawn from '../index.js';
+import nanoSpawn from '../../source/index.js';
 
 await nanoSpawn('node', ['--version'], {stdout: 'inherit'});
