+++ test/fixtures/shebang.js
